Update Containers.md
Difficult one as question suggests there is only one answer however both are correct. If anything, services running slower should be the answer if there is only once correct one as this is the first thing that will likely be experienced. Also, answer suggest that this is the correct checkbox.